Move embed video function to utils
