+++ funnel/assets/js/proposal.js
@@ -1,3 +1,5 @@
+import { Video } from './util';
+
 export const Proposal = {
   init() {
     $('button[name="transition"][value="delete"]').click(function(e) {
@@ -99,41 +101,6 @@ export const Comments = {
       $('#comment-form textarea').focus();
       return false;
     });
-  },
-};
-
-export const Video = {
-  /* Takes argument
-     `videoWrapper`: video container element,
-     'videoUrl': video url
-    Video id is extracted from the video url (extractYoutubeId).
-    The videoID is then used to generate the iframe html.
-    The generated iframe is added to the video container element.
-  */
-  getVideoTypeAndId(url) {
-    const regexMatch = url.match(
-      /(http:|https:|)\/\/(player.|www.)?(vimeo\.com|youtu(be\.com|\.be|be\.googleapis\.com))\/(video\/|embed\/|watch\?v=|v\/)?([A-Za-z0-9._%-]*)(&\S+)?/
-    );
-    let type = '';
-    if (regexMatch[3].indexOf('youtu') > -1) {
-      type = 'youtube';
-    } else if (regexMatch[3].indexOf('vimeo') > -1) {
-      type = 'vimeo';
-    }
-    return {
-      type,
-      videoId: regexMatch[6],
-    };
-  },
-  embedIframe(videoWrapper, videoUrl) {
-    let videoEmbedUrl;
-    const { type, videoId } = this.getVideoTypeAndId(videoUrl);
-    if (type === 'youtube') {
-      videoEmbedUrl = `<iframe src='//www.youtube.com/embed/${videoId}' frameborder='0' allowfullscreen></iframe>`;
-    } else if (type === 'vimeo') {
-      videoEmbedUrl = `<iframe src='https://player.vimeo.com/video/${videoId}' frameborder='0' allowfullscreen></iframe>`;
-    }
-    videoWrapper.innerHTML = videoEmbedUrl;
   },
 };
