--- funnel/assets/js/settings.js
@@ -1,9 +1,0 @@
-import { SaveProject } from './util';
-
-$(() => {
-  window.HasGeek.Settings = function init(saveProjectConfig = '') {
-    if (saveProjectConfig) {
-      SaveProject(saveProjectConfig);
-    }
-  };
-});
