+++ funnel/static/css/app.css
@@ -370,6 +370,36 @@ html.touch .collapsible__body--disable {
   display: none;
 }
 
+.embed-video-wrapper {
+  position: relative;
+  width: 100%;
+  height: 0;
+  padding-bottom: 56%;
+}
+.embed-video-wrapper .video_txt {
+  text-align: center;
+  position: absolute;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  left: 50%;
+}
+.embed-video-wrapper iframe {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  left: 0;
+  top: 0;
+  display: block;
+}
+.embed-video-wrapper .no-video {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 32px;
+  color: #8492a6;
+}
+
 .hg-app .header--fixed {
   bottom: 0;
 }
@@ -1486,7 +1516,6 @@ html.touch .collapsible__body--disable {
 .card__image-wrapper {
   min-height: 180px;
   position: relative;
-  padding-bottom: 8px;
   background-image: linear-gradient(to bottom, #4d5763, #202d3c);
 }
 .card__image-wrapper .card__image__tagline {
@@ -1954,22 +1983,7 @@ html.touch .collapsible__body--disable {
 .project-banner .project-banner__box {
   padding: 0;
 }
-.project-banner .project-banner__box .project-banner__box__16_9 {
-  padding-bottom: 56%;
-  position: relative;
-}
-.project-banner .project-banner__box .project-banner__box__16_9 iframe {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  left: 0;
-  top: 0;
-  display: block;
-}
-.project-banner
-  .project-banner__box
-  .project-banner__box__16_9
-  .project-banner__box__image {
+.project-banner .project-banner__box .project-banner__box__image {
   position: absolute;
   object-fit: cover;
   width: 100%;
@@ -2338,36 +2352,6 @@ html.touch .collapsible__body--disable {
 
 .proposal-details {
   margin-top: 15px;
-}
-
-.proposal-video {
-  position: relative;
-  width: 100%;
-  height: 0;
-  padding-bottom: 56%;
-}
-.proposal-video .video_txt {
-  text-align: center;
-  position: absolute;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  left: 50%;
-}
-.proposal-video iframe {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  left: 0;
-  top: 0;
-  display: block;
-}
-.proposal-video .no-video {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  font-size: 32px;
-  color: #8492a6;
 }
 
 .proposal-video--novideo {
