+++ funnel/static/sass/_card.scss
@@ -91,7 +91,6 @@
 .card__image-wrapper {
   min-height: 180px;
   position: relative;
-  padding-bottom: 8px;
   background-image: linear-gradient(to bottom, #4d5763, #202d3c);
 
   .card__image__tagline {
