+++ funnel/static/sass/_project.scss
@@ -20,24 +20,11 @@
   .project-banner__box {
     padding: 0;
 
-    .project-banner__box__16_9 {
-      padding-bottom: 56%;
-      position: relative;
-
-      iframe {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        left: 0;
-        top: 0;
-        display: block;
-      }
-      .project-banner__box__image {
-        position: absolute;
-        object-fit: cover;
-        width: 100%;
-        height: 100%;
-      }
+    .project-banner__box__image {
+      position: absolute;
+      object-fit: cover;
+      width: 100%;
+      height: 100%;
     }
     .project-banner__headline {
       padding: 0 15px;
