+++ funnel/templates/project.html.jinja2
@@ -221,6 +221,7 @@
 {% endblock %}
 
 {% block footerscripts %}
+  <script src="{{ url_for('static', filename=asset_path('project_header')) }}" type="text/javascript"></script>
   <script src="{{ url_for('static', filename=asset_path('project')) }}" type="text/javascript"></script>
   <script type="text/javascript">
     $(function() {
@@ -248,11 +249,13 @@
         };
       {%- endif %}
 
-      projectConfig.saveProjectConfig = {
+      var saveProjectConfig = {
         formId: 'save-form',
-        postUrl: "{{ project.url_for('save') }}"
-      }
+        postUrl: {{ project.url_for('save')|tojson }}
+      };
 
+
+      window.HasGeek.ProjectHeaderInit(saveProjectConfig);
       window.HasGeek.ProjectInit(projectConfig);
 
     });
