+++ funnel/templates/proposal.html.jinja2
@@ -76,7 +76,7 @@
       <div class="mui--clearfix">
         <div class="proposal-details proposal-details--left">
           {% if proposal.preview_video.url %}
-            <div class="proposal-video mui--bg-primary-light" data-href="{{ proposal.preview_video }}" id="js-embed">
+            <div class="proposal-video mui--bg-primary-light" data-video-src="{{ proposal.preview_video }}" id="js-embed">
               <p class="video_txt">
                 {% assets "fa5-sprite" %}<svg class="fa5-icon fa5-icon--display1 fa5--align-baseline" aria-hidden="true" role="img"><use xlink:href="{{ ASSET_URL }}#video"></use></svg>{% endassets %}<br>
                 <a href="{{ proposal.preview_video }}" target="_blank" class="mui--text-title">{% trans %}Preview video{% endtrans %}</a>
@@ -279,7 +279,7 @@
     proposalConfig.pageUrl = {{ request.base_url|tojson }};
     proposalConfig.videoWrapper = document.getElementById('js-embed');
     if(proposalConfig.videoWrapper) {
-      proposalConfig.videoUrl = proposalConfig.videoWrapper.getAttribute('data-href');
+      proposalConfig.videoUrl = proposalConfig.videoWrapper.getAttribute('data-video-src');
     }
     HasGeek.ProposalInit(proposalConfig);
